refactor: add file name config for dev and prod environments
