+++ apps/server/src/utils.rs
@@ -4,9 +4,20 @@ use dirs::data_dir;
 use std::path::PathBuf;
 use tracing::error;
 
-pub fn get_application_support_path() -> PathBuf {
+fn get_db_name() -> String {
+    if cfg!(debug_assertions) {
+        return String::from("skopio_server_test.db");
+    } else {
+        return String::from("skopio_server.db");
+    }
+}
+
+pub fn get_db_path() -> PathBuf {
     let base_dir = data_dir().unwrap_or_else(|| PathBuf::from("."));
-    base_dir.join("Skopio").join("skopio.db")
+    base_dir
+        .join("com.samwahome.skopio")
+        .join("server")
+        .join(get_db_name())
 }
 
 pub fn error_response<E: std::fmt::Display>(err: E) -> (StatusCode, Json<String>) {
